Restore ui/shared/db/scale-options
