+++ src/ui/shared/db/index.ts
@@ -4,3 +4,4 @@ export * from "./creator-form";
 export * from "./selector";
 export * from "./endpoints-detail";
 export * from "./credential";
+export * from "./scale-options";
